Picker on MainPickers.java and activity_main_pickers.xml
