+++ app/src/main/java/com/example/activitiesbutton/MainActivity.java
@@ -1,52 +1,24 @@
 package com.example.activitiesbutton;
 
-import android.annotation.SuppressLint;
-import android.app.DatePickerDialog;
 import android.os.Bundle;
-import android.widget.Button;
-import android.widget.DatePicker;
-import android.widget.TextView;
 
+import androidx.activity.EdgeToEdge;
 import androidx.appcompat.app.AppCompatActivity;
-
-import java.util.Calendar;
+import androidx.core.graphics.Insets;
+import androidx.core.view.ViewCompat;
+import androidx.core.view.WindowInsetsCompat;
 
 public class MainActivity extends AppCompatActivity {
-
-    private TextView tvSelectedDate;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
+        EdgeToEdge.enable(this);
         setContentView(R.layout.activity_main);
-
-        // Initialize views
-        Button btnPickDate = findViewById(R.id.btn_pick_date);
-        tvSelectedDate = findViewById(R.id.tv_selected_date);
-
-        // Set up button click listener
-        btnPickDate.setOnClickListener(v -> showDatePickerDialog());
-    }
-
-    private void showDatePickerDialog() {
-        // Get the current date
-        Calendar calendar = Calendar.getInstance();
-        int year = calendar.get(Calendar.YEAR);
-        int month = calendar.get(Calendar.MONTH);
-        int day = calendar.get(Calendar.DAY_OF_MONTH);
-
-        // Create and show DatePickerDialog
-        DatePickerDialog datePickerDialog = new DatePickerDialog(
-                MainActivity.this,
-                (DatePicker view, int selectedYear, int selectedMonth, int selectedDay) -> {
-                    // Format and display the selected date using string resources
-                    String formattedDate = selectedDay + "/" + (selectedMonth + 1) + "/" + selectedYear;
-                    @SuppressLint({"StringFormatInvalid", "LocalSuppress"}) String dateString = getString(R.string.selected_date, formattedDate);
-                    tvSelectedDate.setText(dateString);
-                },
-                year, month, day
-        );
-
-        datePickerDialog.show();
+        ViewCompat.setOnApplyWindowInsetsListener(findViewById(R.id.main), (v, insets) -> {
+            Insets systemBars = insets.getInsets(WindowInsetsCompat.Type.systemBars());
+            v.setPadding(systemBars.left, systemBars.top, systemBars.right, systemBars.bottom);
+            return insets;
+        });
     }
 }
